Attempt 2 at separated stripsearch_link
This time it actually works and produces small files the way it should.

Currently it still uses the first strip it finds of the 3 faces of each
triangle, like it did before. For some reason I can't figure out, using
the longest strip instead results in holes in the model.